requirements.txt
twisted
qtpy
pyqt5
vext
vext.gi
pyqtwebengine
pywebview
websockets